==> 01_preprocess.ipynb <==
# Notebook for DSW Telkomsel 2023 - Party Parrot

import os
import re
import duckdb
import numpy as np
import pandas as pd

from fuzzywuzzy import fuzz
from tqdm import tqdm

%load_ext autoreload
%autoreload 2

# Read Data

df_name = pd.read_excel('datasets/raw/product_name.xlsx')
catalog = pd.read_excel('datasets/raw/product_catalog.xlsx')

df_name.rename(columns={'Product Name': 'product_name'}, inplace=True)
catalog.rename(columns={
    'Product SKU': 'product_sku',
    'Brand': 'brand',
    'Type': 'type',
    'Formula': 'formula'
}, inplace=True)

df_name = df_name.dropna()
df_name

catalog

d = {
    'basf': {'nitrophoska', 'entec', 'brucite', 'cantik', 'calcinit', 'perfect'},
    'dgw/hextar': {'mroph', 'morph', 'hx', 'hxas', 'gold', 'dgw', 'cockhead', 'booster', 'compaction'},
    'laoying': {'mestikali', 'mestifos', 'patenbiru', 'mestikisrit', 'bomber', 'mestical',
                'mestifoska', 'emcote', 'patenhijo', 'mestac', 'mesti', 'evermax', 'borate', 'lao ying', 'loying'},
    'mahkota': {'borat', 'peru', 'mahkota', 'kieserite', 'triple', 'mesir'},
    'mutiara': {'merokemkp', 'sprinter', 'partner', 'professional', 'meroketsp', 'merokesop',
                'kali', 'flex', 'merokeza', 'butir', 'merokemap', 'grower', 'merokemop',
                'meroke', 'merokekkb', 'karate', 'merokecpn', 'mutiara', 'merokerock',
                'boroni', 'suburkali'},
    'paktani': {'saprodap', 'holland', 'phosgro', 'urecote', 'magnesium', 'padi', 'fertikali',
                'magnum', 'sawit', 'kristalon', 'fertiphos', 'merah', 'tani', 'fertila',
                'sulfate', 'fertilla', 'singkong', 'qrop', 'ultradap'},
    'pihc': {'jeranti', 'kujang', 'solution', 'petro', 'petrofert', 'nitrat', 'pusri', 'npk',
             'mikroba', 'phosphate', 'phonska', 'polivit', 'rock', 'fertigres', 'nitrea', 
             'coating', 'nitroku', 'pelangi', 'biofertil', 'agro', 'kebomas', 'nitralite',
             'niphos', 'ningrat', 'kapur', 'phosgreen', 'pertanian', 'nitroska'},
    'tawon': {'tawon', 'nitroplus', 'vrea', 'astiva'},
    'yara': {'yaraliva', 'yaravita', 'krista', 'winner', 'yaravera', 'unik', 'palmae',
             'yaratera', 'calcinit', 'faster', 'yaramila', 'tropicote', 'complex',
             'pholate', 'nitrabor', 'sulfate', 'ammonium'}
}

d = {k: sorted(list(d[k] | {k})) for k in d}

# convert to flattened Pandas Series
d = pd.Series(d).explode().reset_index()
d.rename(columns={'index': 'brand', 0: 'common_token'}, inplace=True)
d.to_csv('datasets/processed/common_tokens_per_brand.tsv', sep='\t', index=False)

common_tokens = pd.read_csv('datasets/processed/common_tokens_per_brand.tsv', sep='\t', index_col=[0])
common_tokens

# Preprocessing

def is_containing_non_alphanumeric(string):
    return not bool(re.search(r'[^a-zA-Z0-9 ]', string))

df_name['is_only_alphanum'] = df_name['product_name'].apply(is_containing_non_alphanumeric)
catalog['is_only_alphanum'] = catalog['product_sku'].apply(is_containing_non_alphanumeric)

s = '250'
s = re.sub(r'(\d{2})', r'\1 ', s)
s

def extract_formula(s):
    match = re.search(r'\d+\.?,?\d*x\d+\.?,?\d*x+\d+\.?,?\d*', s)
    if match:
        return match.group()
    return np.nan


def give_spaces_before_and_after_punct(s):
    return re.sub(r'([.,\-()])', r' \1 ', s)

s = 'abc,.-()'
give_spaces_before_and_after_punct(s)

def clean(s):
    s = s.lower()

    # give spaces before and after the special characters, that are not number or decimal number
    # e.g. 'abc+c' to 'abc + c', but '12.9' not to '12 . 9'
    s = re.sub(r'([^\d\.,-]+)', r' \1 ', s)
    s = re.sub(r'\s+', ' ', s) # remove extra spaces once again

    # separate value and units like "100ml" to "100 ml" or "50g" to "50 g"
    # that are not wrapped in <...> bracket
    s = re.sub(r'(\d+)([a-zA-Z]+)', r'\1 \2', s)

    # make chemical formula separator uniform to 'x', i.e. '4.5 - 3.6 - 2.1' to '4.5x3.6x2.1'
    s = re.sub(r'(\d+\.?,?\d*)[×x\+,\. |-]+(\d+\.?,?\d*)[×x\+,\. |-]+(\d+\.?,?\d*)', r'\1x\2x\3', s)

    s = re.sub(r'\s+', ' ', s) # remove extra spaces once again
    
    return s

catalog['clean_sku'] = catalog.product_sku.apply(clean)
catalog['brand'] = catalog.brand.str.lower().str.replace(' ', '')

unique_brand = sorted(catalog.brand.unique())
df_name['clean_name'] = df_name.product_name.apply(clean)

df_name[df_name.is_only_alphanum == False].loc[5200:5400]

catalog[catalog.is_only_alphanum == False].head(20)

df_name.to_csv('datasets/processed/product_name.tsv', sep='\t', index=False)
catalog.to_csv('datasets/processed/product_catalog.tsv', sep='\t', index=False)
