==> 02_postanalysis.ipynb <==
# Notebook for DSW Telkomsel 2023 - Party Parrot

import os
import re
import duckdb
import numpy as np
import pandas as pd

from fuzzywuzzy import fuzz
from tqdm import tqdm

%load_ext autoreload
%autoreload 2

#  Exploration

df_name = pd.read_csv('datasets/processed/product_name.tsv', sep='\t')
catalog = pd.read_csv('datasets/processed/product_catalog.tsv', sep='\t')

name_counts = df_name.clean_name.value_counts()
name_counts.to_frame()

joined = pd.merge(name_counts.to_frame().reset_index(), catalog[['sku_id', 'clean_sku']], how='cross')
joined

def is_name_token_present_in_sku(name, sku) -> bool:
    name_tokens = name.split()
    sku_tokens = sku.split()
    for token in name_tokens:
        if (len(token) <= 2 and token != 'za') or token in ['plus']:
            continue
        if token in sku_tokens:
            return True
    return False

duckdb.remove_function("is_name_token_present_in_sku")
duckdb.create_function("is_name_token_present_in_sku", is_name_token_present_in_sku)

hard_match = duckdb.query(
    """
    WITH cte AS (
        SELECT
            clean_name,
            clean_sku,
            sku_id,
            is_name_token_present_in_sku(clean_name, clean_sku) as is_token_in_name_in_sku
        FROM
            joined
    )

    SELECT
        *
    FROM
        cte
    WHERE is_token_in_name_in_sku
    """
).to_df()

np.random.seed(7)
hard_match.sample(20)

df_name[~df_name.clean_name.isin(hard_match.clean_name.unique())]

all_tokens = []
df_name.clean_name.apply(lambda x: all_tokens.extend(x.split(' ')))
all_tokens = [token for token in all_tokens if token.isalpha() and len(token) > 2]
all_tokens = pd.Series(all_tokens)
token_counts = all_tokens.value_counts()
token_counts.head(20)

name_counts[(name_counts <= 1) & (name_counts.index.str.len() <= 2)].sort_index()

# Post-analysis

final_res = pd.read_csv('datasets/processed/result_fuzzy.tsv', sep='\t')
print(set(range(44002)) - set(final_res.product_id.tolist()))
# final_res.to_csv('datasets/processed/final_result.tsv', sep='\t', index=False)
final_res

final_res.loc[
    final_res['is_name_token_present_in_sku_fuzzy'] & (final_res['fuzzy_ratio'] >= 20),
    ['possible_brand', 'clean_name', 'result_clean_sku_lev', 'result_clean_sku_fuzzy', 'fuzzy_ratio', 'cnt_common_tokens_fuzzy']
]. \
    sort_values('fuzzy_ratio', ascending=True). \
    head(20)

cond1 = (
    False
    | (final_res.is_name_token_present_in_sku_fuzzy & (final_res.fuzzy_ratio >= 28))
    | ((final_res.cnt_common_tokens_fuzzy >= 2) & (final_res.fuzzy_ratio >= 28))
    | (final_res.possible_brand.notnull() & (final_res.fuzzy_ratio >= 55))
    | (final_res.possible_brand.isnull()  & (final_res.fuzzy_ratio >= 61))
    | (final_res['lev_dist_fuzzy'].between(0, 2))
    | (final_res['lev_dist_lev'].between(0, 2))

)
print(final_res[cond1].shape)

final_res.loc[
    cond1,
    ['clean_name', 'result_clean_sku_lev', 'result_clean_sku_fuzzy', 'fuzzy_ratio', 'lev_dist_lev', 'lev_dist_fuzzy', 'cnt_common_tokens_fuzzy']
] \
  .sort_values('fuzzy_ratio') \
  .tail(30)

low_fuzzy = final_res[final_res.fuzzy_ratio.between(0, 40)]
low_fuzzy.sort_values(['fuzzy_ratio', 'cnt_common_tokens_fuzzy', 'lev_dist_fuzzy'])

final_res.loc[
    final_res.fuzzy_ratio.between(70, 80),
    ['product_id', 'clean_name', 'result_clean_sku_lev', 'result_clean_sku_fuzzy', 'fuzzy_ratio']
] \
    .sort_values('fuzzy_ratio') \
    .iloc[list(range(10)) + list(range(-10, 0))]

final_res.loc[
    ~cond1 & (final_res.fuzzy_ratio.between(61, 62)),
    ['product_id', 'clean_name', 'result_clean_sku_lev', 'result_clean_sku_fuzzy', 'fuzzy_ratio']
] \
    .sort_values('fuzzy_ratio') \
    .iloc[list(range(10)) + list(range(-10, 0))]
