==> 03_finalize.ipynb <==
import os
import re
import duckdb
import numpy as np
import pandas as pd

from fuzzywuzzy import fuzz
from tqdm import tqdm

%load_ext autoreload
%autoreload 2

result_fuzzy_raw = pd.read_csv('datasets/processed/result_fuzzy.tsv', sep='\t')
result_fuzzy_raw.head()

result_fuzzy = duckdb.query(
    '''
    SELECT
        product_id,
        product_name,
        clean_name_alphanum,
        result_sku_lev,
        result_sku_fuzzy,
        result_sku_id_lev,
        result_sku_id_fuzzy,
        fuzzy_ratio,
        lev_dist_lev AS lev_dist_int
    FROM
        result_fuzzy_raw
    '''
).to_df()
result_fuzzy.head()

result_lev = pd.read_csv('datasets/Similarity Result - result_v2_product_id.csv', sep=',')
result_lev = result_lev.rename(columns={
    'B_product_id': 'product_id',
    'B_string_raw': 'product_name',
    'Best_A_strings_raw': 'sku_name',
    'string_similarity': 'lev_distance'
})
result_lev = result_lev.loc[
    result_lev['lev_distance'] >= 0.59,
    ['product_id', 'product_name', 'sku_name', 'lev_distance']
]

# dedup with duckdb by lev_distance desc
result_lev = duckdb.query(
    '''
    SELECT
        product_id,
        product_name,
        sku_name,
        lev_distance,
    FROM
        result_lev
    QUALIFY
        ROW_NUMBER() OVER (
            PARTITION BY
                product_id
            ORDER BY
                lev_distance DESC
        ) = 1
    ORDER BY
        product_id
    '''
).to_df()

result_lev

# Final Result Post-processing

final_res = duckdb.query(
    '''
    SELECT
        f.product_id,
        f.product_name,
        f.clean_name_alphanum,
        f.result_sku_lev AS result_sku_lev_int,
        f.result_sku_fuzzy,
        f.result_sku_id_lev AS result_sku_id_lev_int,
        f.result_sku_id_fuzzy,
        f.fuzzy_ratio,
        f.lev_dist_int,
        l.product_id AS product_id_lev,
        l.sku_name AS result_sku_lev_ratio,
        l.lev_distance AS lev_dist_ratio
    FROM
        result_fuzzy AS f
    LEFT JOIN
        result_lev AS l
    ON
        f.product_id = l.product_id
    ORDER BY
        f.product_id
    '''
).to_df()

final_res

def is_match_naive_regex(s):
    pat = "beras|ayam|ikan|tapetool|remasil|pasir|megastar|pajero|kastro|menir|" \
        "rapia|top asia|great|greta|emrat|koret|obor|tomat|destan|betras|saporo|" \
        "hadroh|manohara|preza|pare valera|santika|doran|bara|posatbona|apuri|" \
        "toram|bolt|cream|angin|star|petir|minyak|catchoize|lenovo|bela ?rusia|" \
        "^elang |borax"
    return (
        re.search(pat, s, re.IGNORECASE) is not None
        or (len(s) <= 2 and s != 'za')
    )

final_res['is_match_naive_regex'] = final_res['clean_name_alphanum'].apply(is_match_naive_regex)
final_res.head()

# post-processing

final_res = duckdb.query(
    '''
    SELECT
        product_id,
        product_name,
        clean_name_alphanum,

        CASE
            WHEN is_match_naive_regex THEN 'Others'
            WHEN fuzzy_ratio < 40 THEN 'Others'
            WHEN lev_dist_int <= 2 THEN result_sku_lev_int
            WHEN product_id_lev IS NOT NULL THEN result_sku_lev_ratio
            ELSE result_sku_fuzzy
        END AS result_sku,

        CASE
            WHEN is_match_naive_regex THEN 'Others'
            WHEN fuzzy_ratio < 40 THEN 'Others'
            WHEN lev_dist_int <= 2 THEN 'levenshtein'
            WHEN product_id_lev IS NOT NULL THEN 'levenshtein'
            ELSE 'fuzzy'
        END AS method_label,

        fuzzy_ratio,
        is_match_naive_regex,
        lev_dist_int,
        lev_dist_ratio,
    FROM
        final_res
    ORDER BY
        product_id
    '''
).to_df()

final_res

catalog = pd.read_csv('datasets/processed/product_catalog.tsv', sep='\t')

# remove duplicates by product_sku
catalog = catalog.drop_duplicates(subset=['product_sku'])
print(catalog.shape)

catalog.head()

final_res = duckdb.query(
    '''
    SELECT
        product_id,
        product_name,
        clean_name_alphanum AS product_clean_name,
        result_sku,
        c.brand AS result_brand,
        c.type AS result_type,
        c.formula AS result_formula,
        method_label,
        fuzzy_ratio,
        is_match_naive_regex,
        lev_dist_int,
        lev_dist_ratio
    FROM
        final_res AS f
    LEFT JOIN
        catalog AS c
    ON
        f.result_sku = c.product_sku
    ORDER BY
        f.product_id
    '''
).to_df()
final_res.shape

final_res.to_csv('datasets/final_result.tsv', sep='\t', index=False)
final_res.head(10)

final_res.method_label.value_counts()

final_res[['method_label', 'result_sku']].value_counts()
